Update certificate loaded path, setup package, and bug
Some debug messages were added in order to validate the certificate file
path loading process. The package needed to complete this operation was
added to setup.by.

There was a mistake in function Search_region regarding the name of the
variables and there's still another bug in the Process_data funtion call
inside this function, for which I added a comment indicating the error.